refactor: 패키지 구조 변경
- factory가 domain과 controller 등에 섞여있었는데 분리함
