+++ src/main/java/lotto/Application.java
@@ -1,7 +1,7 @@
 package lotto;
 
 import lotto.controller.LottoController;
-import lotto.controller.LottoControllerFactory;
+import lotto.factory.LottoControllerFactory;
 
 public class Application {
     public static void main(String[] args) {
