+++ src/main/java/lotto/factory/LottoControllerFactory.java
@@ -1,4 +1,6 @@
-package lotto.controller;
+package lotto.factory;
+
+import lotto.controller.LottoController;
 
 public class LottoControllerFactory {
     public LottoController lottoController() {
