+++ src/main/java/lotto/factory/LottoFactory.java
@@ -1,8 +1,9 @@
-package lotto.domain;
+package lotto.factory;
 
 import java.util.ArrayList;
 import java.util.List;
 import camp.nextstep.edu.missionutils.Randoms;
+import lotto.domain.Lotto;
 
 public class LottoFactory {
     private static final int START = 1;
